requirements.txt
fastapi>=0.100
uvicorn[standard]==0.22.0
pydantic>=2.0.0,<3.0.0
pymongo==4.10.1
transformers==4.34.0
torch>=1.13.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
# google-generativeai==1.3.0
google-genai==0.5.0
numpy==1.25.2
scikit-learn==1.3.2
